tools: Clean up and simplify download metadata generation
