+++ .github/workflows/build.yml
@@ -336,13 +336,10 @@ jobs:
       id: generate_artifact_metadata
       # skip deploying Ubuntu builds to downloads.mixxx.org because these are deployed to the PPA
       if: github.event_name == 'push'
-      run: python3 tools/generate_artifact_metadata.py ${{ matrix.artifacts_path }}
+      run: python3 tools/generate_download_metadata.py artifact ${{ matrix.artifacts_path }} "${{ matrix.artifacts_slug }}"
       env:
-        OS: ${{ runner.os }}
-        OS_IMAGE: ${{ matrix.os }}
         DEPLOY_BASEURL: "https://downloads.mixxx.org"
-        DEPLOY_DESTPATH: "builds/{git_branch}/{os_name}"
-        ARTIFACT_SLUG: ${{ matrix.artifacts_slug }}
+        DESTDIR: builds/{git_branch}/${{ runner.os }}
 
     - name: "[macOS/Windows] Upload build to downloads.mixxx.org"
       # skip deploying Ubuntu builds to downloads.mixxx.org because these are deployed to the PPA
@@ -382,11 +379,11 @@ jobs:
 
     - name: "Collect Artifacts Metadata & Write Manifest"
       if: github.event_name == 'push' && env.SSH_PASSWORD != null
-      run: python3 tools/collect_artifacts_metadata.py
+      run: python3 tools/generate_download_metadata.py manifest
       env:
         JOB_DATA: ${{ toJSON(needs.build) }}
         MANIFEST_URL: "https://downloads.mixxx.org/builds/{git_branch}/manifest.json"
-        DEPLOY_DESTPATH: "builds/{git_branch}"
+        DESTDIR: "builds/{git_branch}"
         SSH_PASSWORD: ${{ secrets.DOWNLOADS_HOSTGATOR_DOT_MIXXX_DOT_ORG_KEY_PASSWORD }}
 
     - name: "Deploy Manifest"
